backend/medicar/requirements.txt
Django==3.1.1
djangorestframework==3.11.1
django-filter==2.3.0

# Debug
django-debug-toolbar==3.0

# MySQL
mysqlclient==2.0.1

